fix: extracted component-data into separate module
