+++ frontend/Components/Guest/utils/processData.js
@@ -1,0 +1,29 @@
+const processes = [
+  {
+    title: "Sign up for an account",
+    description:
+      'You can easily sign up via "Facebook" or "Google" and be a member of NoteIT.',
+  },
+  {
+    title: "Get a list of available notes",
+    description:
+      "We will send you a list of notes based on your faculty and semester.",
+  },
+  {
+    title: "Browse the notes",
+    description:
+      "Easily view and read notes on up to three different device(s), using the same account.",
+  },
+  {
+    title: "Save the notes",
+    description:
+      'You can "save" the notes into your account to have easy access later on.',
+  },
+  {
+    title: "Ask questions",
+    description:
+      'Clear your confusions by chatting with the "note contributor" themselves.',
+  },
+];
+
+export default processes;
